feat: moves list sessions to abstract sessions module
