+++ apps/api/src/modules/archived-sessions/infrastructure/queries/list-archived-nomination-sessions.query.ts
@@ -9,10 +9,12 @@ import { createPaginatedZodDto, paginate, Pagination } from 'src/modules/framewo
 import { Sortable } from 'src/modules/framework/sorting';
 import { FormationEnum } from 'src/modules/shared/formation.enum';
 import { prismaFormationEnumToFormationEnum } from 'src/modules/shared/mappers/formation.mapper';
-import { prismaTypeDeSaisineEnumToTypeDeSaisine } from 'src/modules/shared/mappers/type-de-saisine-enum.mapper';
+import {
+  prismaTypeDeSaisineEnumToTypeDeSaisine,
+  typeDeSaisineToPrismaTypeDeSaisineEnum,
+} from 'src/modules/shared/mappers/type-de-saisine-enum.mapper';
 import { TypeDeSaisineEnum } from 'src/modules/shared/type-de-saisine.enum';
-import { dateOnlyJsonSchema } from 'src/utils/date-only';
-import { DateOnly } from 'src/utils/date-only';
+import { DateOnly, dateOnlyJsonSchema } from 'src/utils/date-only';
 
 const SESSION_STATUSES = ['TO_VALIDATE', 'READY'] as const;
 type SessionStatus = (typeof SESSION_STATUSES)[number];
@@ -32,7 +34,7 @@ export class ListArchivedNominationSessionsQuery {
     const where: Prisma.SessionWhereInput = {
       deletedAt: null,
       archivedAt: { not: null },
-      typeDeSaisine: query.typeDeSaisine,
+      typeDeSaisine: typeDeSaisineToPrismaTypeDeSaisineEnum(query.typeDeSaisine),
       ...(query.formations?.length && {
         formation: { in: [...query.formations] },
       }),
