+++ apps/api/src/modules/root.module.ts
@@ -10,6 +10,7 @@ import { MaintenanceModule } from './maintenance/maintenance.module';
 import { MembersModule } from './members';
 import { ObservationModule } from './observation/observation.module';
 import { ReportModule } from './report/report.module';
+import { AbstractSessionModule } from './session/abstract/abstract-session.module';
 import { TransparenceModule } from './session/transparence/transparence.module';
 import { SimpleAuthModule } from './simple-auth';
 
@@ -18,6 +19,7 @@ import { SimpleAuthModule } from './simple-auth';
     SimpleAuthModule,
     ReportModule,
     IngestModule,
+    AbstractSessionModule,
     TransparenceModule,
     ArchivedSessionsModule,
     MembersModule,
