+++ apps/api/src/modules/shared/mappers/type-de-saisine-enum.mapper.ts
@@ -17,7 +17,7 @@ export function typeDeSaisineToPrismaTypeDeSaisineEnum(value: TypeDeSaisineEnum)
   switch (value) {
     case 'TRANSPARENCE_GDS':
       return 'TRANSPARENCE_GDS';
-    case TypeDeSaisine.MTT:
+    case 'MTT':
       return 'MTT';
     default:
       return assertNever(value);
